using uint32_t and uint64_t causes bad casting and the code to sort threads is not calculating times correctly. This patch fixes both of these in the actor code
git-svn-id: svn://svn.berlios.de/open-iscsi@743 d7303112-9cec-0310-bdd2-e83a94d6c2b6

+++ usr/actor.h
@@ -36,8 +36,8 @@ typedef struct actor {
     actor_state_e state;
     void *data;
     void (*callback)(void * );
-    uint32_t scheduled_at;
-    uint32_t ttschedule;
+    uint64_t scheduled_at;
+    uint64_t ttschedule;
 } actor_t;
 
 extern void actor_new(actor_t *thread, void (*callback)(void *), void * data);
